+++ docs/examples/datefield/error.tsx
@@ -3,37 +3,39 @@ import { Box, Flex } from 'gestalt';
 import { DateField } from 'gestalt-datepicker';
 
 export default function Example() {
-  const [dateValue, setDateValue] = useState<Date | null>(null);
-  const [errorText, setErrorText] = useState<string | null | undefined>(null);
+  const [dateValue1] = useState<Date | null>(null);
+  const [dateValue2] = useState<Date | null>(null);
 
   return (
-    <Flex alignItems="center" gap={4} height="100%" justifyContent="center" width="100%">
-      <Box width={400}>
+    <Box padding={8} width="100%">
+      <Flex direction="column" gap={6} width="100%">
         <DateField
           disableRange="disableFuture"
-          errorMessage={errorText || undefined}
+          errorMessage="Please, select a valid birth date"
           helperText="Enter your date of birth"
-          id="errorExample"
+          id="error-datefield-md"
           label="Date of birth"
           name="bday_datefield"
-          onChange={({ value }) => {
-            setDateValue(value);
-          }}
-          onClearInput={() => {
-            setErrorText(null);
-            setDateValue(null);
-          }}
-          onError={({ errorMessage, value }) => {
-            const date = value ? new Date(value) : null;
-
-            if (errorMessage === 'invalidDate') return;
-            if (errorMessage === 'disableFuture' || (date && date.getFullYear() === 1))
-              setErrorText('Please, select a valid birth date');
-            if (date && date.getFullYear() > 1) setErrorText(null);
-          }}
-          value={dateValue}
+          onChange={() => {}}
+          onClearInput={() => {}}
+          onError={() => {}}
+          size="md"
+          value={dateValue1}
         />
-      </Box>
-    </Flex>
+        <DateField
+          disableRange="disableFuture"
+          errorMessage="Please, select a valid birth date"
+          helperText="Enter your date of birth"
+          id="error-datefield-lg"
+          label="Date of birth"
+          name="bday_datefield"
+          onChange={() => {}}
+          onClearInput={() => {}}
+          onError={() => {}}
+          size="lg"
+          value={dateValue2}
+        />
+      </Flex>
+    </Box>
   );
 }