+++ docs/examples/numberfield/error.tsx
@@ -2,16 +2,18 @@ import { useState } from 'react';
 import { Box, Flex, NumberField } from 'gestalt';
 
 export default function Example() {
-  const [input1text, setInput1Text] = useState<number | undefined>(0);
-  const [input2text, setInput2Text] = useState<number | undefined>(0);
-  const [input3text, setInput3Text] = useState<number | undefined>(0);
+  const [input1text, setInput1Text] = useState<number | undefined>(undefined);
+  const [input2text, setInput2Text] = useState<number | undefined>(undefined);
+  const [input3text, setInput3Text] = useState<number | undefined>(undefined);
 
   return (
     <Box padding={8} width="100%">
       <Flex direction="column" gap={6} width="100%">
         <NumberField
-          disabled
-          id="disabled-example-sm"
+          errorMessage={
+            input1text === null || input1text === undefined ? 'You must enter a number' : null
+          }
+          id="error-example-sm"
           label="Number of widgets"
           onChange={({ value }) => setInput1Text(value)}
           placeholder="Please enter the number of widgets"
@@ -19,8 +21,10 @@ export default function Example() {
           value={input1text}
         />
         <NumberField
-          disabled
-          id="disabled-example-md"
+          errorMessage={
+            input2text === null || input2text === undefined ? 'You must enter a number' : null
+          }
+          id="error-example-md"
           label="Number of widgets"
           onChange={({ value }) => setInput2Text(value)}
           placeholder="Please enter the number of widgets"
@@ -28,8 +32,10 @@ export default function Example() {
           value={input2text}
         />
         <NumberField
-          disabled
-          id="disabled-example-lg"
+          errorMessage={
+            input3text === null || input3text === undefined ? 'You must enter a number' : null
+          }
+          id="error-example-lg"
           label="Number of widgets"
           onChange={({ value }) => setInput3Text(value)}
           placeholder="Please enter the number of widgets"